Move top level Go code for Go Sync into packages folder
