+++ packages/gosync/errors.go
