+++ packages/gosync/mock_Adapter_test.go
