+++ packages/slack/go.mod
@@ -3,7 +3,7 @@ module github.com/ovotech/go-sync/packages/slack
 go 1.18
 
 require (
-	github.com/ovotech/go-sync v0.0.0
+	github.com/ovotech/go-sync/packages/gosync v0.0.0
 	github.com/slack-go/slack v0.12.1
 	github.com/stretchr/testify v1.8.2
 )
@@ -17,4 +17,4 @@ require (
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
 
-replace github.com/ovotech/go-sync v0.0.0 => ../../
+replace github.com/ovotech/go-sync/packages/gosync v0.0.0 => ../gosync
