+++ packages/terraformcloud/go.mod
@@ -4,7 +4,7 @@ go 1.18
 
 require (
 	github.com/hashicorp/go-tfe v1.18.0
-	github.com/ovotech/go-sync v0.0.0
+	github.com/ovotech/go-sync/packages/gosync v0.0.0
 	github.com/stretchr/testify v1.8.2
 )
 
@@ -23,4 +23,4 @@ require (
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
 
-replace github.com/ovotech/go-sync v0.0.0 => ../../
+replace github.com/ovotech/go-sync/packages/gosync v0.0.0 => ../gosync
